# smartfactory-angular-ui/src/app/services/api/model/jsonPatch.ts
/**
 * Smart Factory IoT
 * Smart Factory IoT
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * A JSONPatch document as defined by RFC 6902.
 */
export interface JsonPatch { 
    /**
     * The operation to be performed
     */
    op: string;
    /**
     * A JSON-Pointer
     */
    path: string;
    /**
     * The value to be used within the operations.
     */
    value: any;
    /**
     * A string containing a JSON Pointer value.
     */
    from?: string;
}