# smartfactory-angular-ui/src/app/services/api/model/mail.ts
/**
 * Smart Factory IoT
 * Smart Factory IoT
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Mail { 
    mailFrom?: string;
    mailTo?: string;
    mailCC?: string;
    mailBcc?: string;
    mailSubject?: string;
    mailContent?: string;
    contentType?: string;
}