# smartfactory-angular-ui/src/app/services/api/model/pagination.ts
/**
 * Smart Factory IoT
 * Smart Factory IoT
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Data } from './data';

export interface Pagination {
  data?: Array<Data>;
  currentPage?: number;
  totalItems?: number;
  totalPages?: number;
  pageSize?: number;
}