# smartfactory-angular-ui/src/app/services/api/model/patchBody.ts
/**
 * Smart Factory IoT
 * Smart Factory IoT
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { JsonPatch } from './jsonPatch';

export interface PatchBody extends Array<JsonPatch> { 
}